refactor: 이메일 서비스 리팩토링
