+++ notification-service/src/main/java/com/flight_booking/notification_service/application/service/MailSender.java
@@ -1,0 +1,5 @@
+package com.flight_booking.notification_service.application.service;
+
+public interface MailSender {
+  boolean send(String to, String subject, String content);
+}
