+++ notification-service/src/main/java/com/flight_booking/notification_service/presentation/controller/NotificationController.java
@@ -3,9 +3,11 @@ package com.flight_booking.notification_service.presentation.controller;
 import com.flight_booking.notification_service.presentation.dto.NotificationRequest;
 import com.flight_booking.notification_service.presentation.dto.NotificationResponse;
 import com.flight_booking.notification_service.application.service.NotificationService;
+import com.flight_booking.notification_service.global.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -20,33 +22,37 @@ public class NotificationController {
 
   // 알림 생성
   @PostMapping
-  public ResponseEntity<NotificationResponse> createNotification(@RequestBody NotificationRequest request) {
-    return ResponseEntity.ok(notificationService.createNotification(request));
+  public ResponseEntity<ApiResponse<NotificationResponse>> createNotification(@RequestBody NotificationRequest request) {
+    NotificationResponse response = notificationService.createNotification(request);
+    return ResponseEntity.status(HttpStatus.CREATED).body(ApiResponse.ok(response));
   }
 
   // 특정 알림 조회
   @GetMapping("/{id}")
-  public ResponseEntity<NotificationResponse> getNotification(@PathVariable UUID id) {
-    return ResponseEntity.ok(notificationService.getNotificationById(id));
+  public ResponseEntity<ApiResponse<NotificationResponse>> getNotification(@PathVariable UUID id) {
+    NotificationResponse response = notificationService.getNotificationById(id);
+    return ResponseEntity.ok(ApiResponse.ok(response));
   }
 
   // 사용자 알림 목록 조회 (페이징)
   @GetMapping
-  public ResponseEntity<Page<NotificationResponse>> getUserNotifications(
+  public ResponseEntity<ApiResponse<Page<NotificationResponse>>> getUserNotifications(
       @RequestParam Long userId, Pageable pageable) {
-    return ResponseEntity.ok(notificationService.getNotificationsByUserId(userId, pageable));
+    Page<NotificationResponse> response = notificationService.getNotificationsByUserId(userId, pageable);
+    return ResponseEntity.ok(ApiResponse.ok(response));
   }
 
   // 알림 읽음 처리
   @PatchMapping("/{id}/read")
-  public ResponseEntity<NotificationResponse> markAsRead(@PathVariable UUID id) {
-    return ResponseEntity.ok(notificationService.markAsRead(id));
+  public ResponseEntity<ApiResponse<NotificationResponse>> markAsRead(@PathVariable UUID id) {
+    NotificationResponse response = notificationService.markAsRead(id);
+    return ResponseEntity.ok(ApiResponse.ok(response));
   }
 
   // 알림 삭제
   @DeleteMapping("/{id}")
-  public ResponseEntity<String> deleteNotification(@PathVariable UUID id) {
+  public ResponseEntity<ApiResponse<String>> deleteNotification(@PathVariable UUID id) {
     notificationService.deleteNotification(id);
-    return ResponseEntity.ok("Notification deleted successfully.");
+    return ResponseEntity.ok(ApiResponse.ok("Notification deleted successfully."));
   }
 }
